ButtonSettings: Avoid leaving the navbar preference category empty
On devices with hardware keys, toggling on-screen navbar and then returning
to the previous state results in empty preference category which doesn't
look that pretty.

Simply ditch the removal of the preferences belonging to that category
because the latter is already disabled in case the hardware keys are disabled
during onCreate / onPreferenceTreeClick.

Change-Id: Ida8dff23e926a4965c6b865858272082c2653af1